+++ SurgSim/Collision/UnitTests/TriangleMeshPlaneContactCalculationTests.cpp
@@ -165,7 +165,7 @@ TEST(TriangleMeshPlaneContactCalculationTests, UnitTests)
 	}
 	for (int i = 0; i < cubeNumEdges; ++i)
 	{
-		std::array<unsigned int, 2> edgePoints;
+		std::array<size_t, 2> edgePoints;
 		for (int j = 0; j < 2; ++j)
 		{
 			edgePoints[j] = cubeEdges[i][j];
@@ -176,7 +176,7 @@ TEST(TriangleMeshPlaneContactCalculationTests, UnitTests)
 	}
 	for (int i = 0; i < cubeNumTriangles; ++i)
 	{
-		std::array<unsigned int, 3> trianglePoints;
+		std::array<size_t, 3> trianglePoints;
 		for (int j = 0; j < 3; ++j)
 		{
 			trianglePoints[j] = cubeTrianglesCCW[i][j];
